change domain entity
